requirements.txt
loguru==0.6.0
numpy==1.23.5
PyYAML==6.0
